requirements.txt
drum
feedparser
django-avatar
markdown
django-uuslug
gunicorn
django-markdown
python-social-auth
Mezzanine==4.1.0
djangorestframework==3.5.4
django-cors-headers=2.0.2
djangorestframework-jwt
social-auth-app-django
